requirements.txt
# Core dependencies for ambient-wifi-monitor
# Python 3.9+

# Data processing and analysis
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Statistical analysis
statsmodels>=0.14.0

# Data serialization
pyyaml>=6.0

# Logging and utilities
python-dateutil>=2.8.0

# Optional: Future visualization support
# matplotlib>=3.7.0
# plotly>=5.14.0
